♻️ :: Meal, MealType 클래스 분리
